[J:MEP-0] phase-3 review: align package + Koin-module names with nonWear sibling path
Addresses review findings #1 and #2 (package follows path on renames;
projectsDrivingImplModule misnamed for its new module). Both stem from
the previous commit taking the shortcut of keeping imports resolving
via a same-package classpath trick.

- feat/projects/fe/driving/nonWear/src/*/kotlin/.../projects/fe/driving/
  impl -> .../projects/fe/driving/nonwear (git renames across commonMain,
  webMain, nonWebMain).
- package declarations in moved files updated to ....nonwear.*; internal
  imports pointing to other nonWear-hosted symbols (ui, di) follow.
  Imports of parent-hosted symbols (vm layer: ViewModels, UiStates,
  VmStates, Mappers) stay pointing at ....impl.* — those stayed in
  parent.
- Koin module val projectsDrivingImplModule -> projectsDrivingNonWearModule;
  file ProjectsDrivingModule.kt -> ProjectsDrivingNonWearModule.kt plus
  webMain/nonWebMain expect/actual companions. Consumer import in
  FrontendModulesAggregate updated to match.

Compose-resources namespace (snag.feat.projects.fe.driving.nonwear.
generated.resources) is unchanged; the generator already lowercases the
module name segment.
