+++ feat/projects/fe/driving/nonWear/src/commonMain/kotlin/cz/adamec/timotej/snag/projects/fe/driving/nonwear/internal/projectAssignments/ui/ProjectAssignmentsScreen.kt
@@ -1,0 +1,42 @@
+/*
+ * Copyright (c) 2026 Timotej Adamec
+ * SPDX-License-Identifier: MIT
+ *
+ * This file is part of the thesis:
+ * "Multiplatform snagging system with code sharing maximisation"
+ *
+ * Czech Technical University in Prague
+ * Faculty of Information Technology
+ * Department of Software Engineering
+ */
+
+package cz.adamec.timotej.snag.projects.fe.driving.nonwear.internal.projectAssignments.ui
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import cz.adamec.timotej.snag.lib.design.fe.error.ShowSnackbarOnError
+import cz.adamec.timotej.snag.projects.fe.driving.impl.internal.projectAssignments.vm.ProjectAssignmentsViewModel
+import org.koin.compose.viewmodel.koinViewModel
+import org.koin.core.parameter.parametersOf
+import kotlin.uuid.Uuid
+
+@Composable
+internal fun ProjectAssignmentsScreen(
+    projectId: Uuid,
+    onBack: () -> Unit,
+    viewModel: ProjectAssignmentsViewModel = koinViewModel { parametersOf(projectId) },
+) {
+    val state by viewModel.state.collectAsStateWithLifecycle()
+
+    ShowSnackbarOnError(
+        uiErrorsFlow = viewModel.errorsFlow,
+    )
+
+    ProjectAssignmentsContent(
+        state = state,
+        onAssignUser = viewModel::onAssignUser,
+        onRemoveUser = viewModel::onRemoveUser,
+        onBack = onBack,
+    )
+}
